requirements.txt
streamlit==1.35.0
faiss-cpu
sentence-transformers
numpy
scikit-learn
pandas
